Documentation, rename usermem.h to vars.h
