+++ docs/headers/index.rst
@@ -1,7 +1,7 @@
 .. _headers:
 
 CE Toolchain Specific Headers
-===============================
+=============================
 
 These headers are CE specific, and contain various defines and prototypes.
