+++ docs/headers/sys/basicusb.rst
@@ -1,0 +1,20 @@
+.. _basicusb_h:
+
+sys/basicusb.h
+==============
+
+.. code-block:: c
+
+    #include <sys/basicusb.h>
+
+Proper support for USB devices is in development.
+This header does not provide that.
+
+.. contents:: :local:
+   :depth: 3
+
+API Documentation
+-----------------
+
+.. doxygenfile:: sys/basicusb.h
+   :project: CE C/C++ Toolchain
