+++ docs/headers/sys/index.rst
@@ -1,0 +1,17 @@
+.. _sys:
+
+CE Hardware Headers
+===================
+
+These headers are specific to the calculator's hardware, and contain various defines and prototypes.
+
+.. toctree::
+    :maxdepth: 1
+    :glob:
+
+    basicusb
+    lcd
+    power
+    rtc
+    timers
+    util
