+++ docs/headers/sys/power.rst
@@ -1,0 +1,20 @@
+.. _power_h:
+
+sys/power.h
+===========
+
+.. code-block:: c
+
+    #include <sys/power.h>
+
+Power control functions are not directly accessible.
+However, this header provides access to OS routines that perform control functions.
+
+.. contents:: :local:
+   :depth: 3
+
+API Documentation
+-----------------
+
+.. doxygenfile:: sys/power.h
+   :project: CE C/C++ Toolchain
