+++ docs/headers/sys/rtc.rst
@@ -1,13 +1,14 @@
-.. _certc.h:
+.. _rtc.h:
 
-certc.h
+sys/rtc.h
 ======
 
 .. code-block:: c
 
-    #include <certc.h>
+    #include <sys/rtc.h>
 
 This header includes defines for the Real-Time Clock (RTC) used in the CE.
+The toolchain allows both direct access to the hardware registers, and using TI's helper routines.
 
 .. contents:: :local:
    :depth: 3
@@ -15,5 +16,5 @@ This header includes defines for the Real-Time Clock (RTC) used in the CE.
 API Documentation
 -----------------
 
-.. doxygenfile:: certc.h
+.. doxygenfile:: sys/rtc.h
    :project: CE C/C++ Toolchain
