+++ docs/headers/ti/ceos.rst
@@ -1,0 +1,19 @@
+.. _ceos_h:
+
+ceos.h
+======
+
+.. code-block:: c
+
+    #include <ceos.h>
+
+This header includes defines for OS memory and prototypes for OS routines.
+
+.. contents:: :local:
+   :depth: 3
+
+API Documentation
+-----------------
+
+.. doxygenfile:: ceos.h
+   :project: CE C/C++ Toolchain
