+++ docs/headers/ti/index.rst
@@ -1,0 +1,31 @@
+.. _ti:
+
+TI Operating System Headers
+===========================
+
+These headers are specific to functionality provided by the calculator's operating system, and contain various defines and prototypes.
+
+TI used to provide a free SDK for the TI-83 Plus, which is now no longer available.
+However, the TI-84 Plus CE software is directly derived from the TI-83 Plus software,
+so we do have a lot of official, though dated, information.
+This is also why names are inconsistent: sometimes TI's names don't follow a pattern, and other times,
+whoever in the community first discovered something gave it a name following their own logic.
+
+.. toctree::
+    :maxdepth: 1
+    :glob:
+
+    error
+    flags
+    getcsc
+    getkey
+    graphscreen
+    homescreen
+    hooks
+    info
+    python
+    real
+    realcpp
+    tokens
+    ui
+    vars
