+++ src/ce/ti/vars.h
