+++ src/fileioc/fileioc.h
@@ -13,7 +13,7 @@
 #include <stdlib.h>
 #include <stdbool.h>
 #include <stdio.h>
-#include <ti/usermem.h>
+#include <ti/vars.h>
 
 #ifdef __cplusplus
 extern "C" {
